Added Git history comparator (#walker)
Hashwalker is a new tool that allows you to compare two branches or
refspecs with completely different histories to find the commit where
they differ the least.

To use it you will need to add your disparate repositories as remotes on
a single repository on your local machine.  You'll need to have two
checked out branches to compare, as well.  One branch is your target
you'd like to rebase and the other is an upstream you'd like to rebase
upon.

Running the tool will give you a minimal diff that you can rebase your
problem branch upon, so you can begin bringing it up-to-date with
upstream.

See the "Usage" in scripts/hashwalker for more information.