[NEW] make-lsp: Generate clangd command JSON for make targets
Currently my silly-cc project kind of sucks, but I need a way to
generate my compile commands for LSP on Makefile-based and
Autotools-based projects.  This shell script should work most of the
time provided your Makefiles aren't silencing the emission of compile
commands with a prepending dash ("-") on targets that we care about.